--- a/test/dates/io.jl
+++ b/test/dates/io.jl
@@ -67,13 +67,13 @@ b2 = "96/Feb/1"
 b3 = "96/2/15"
 @test_throws ArgumentError Dates.DateTime(b3,f)
 try
-    DateTime("2012-02-20T09:09:3.43i9")
+    Dates.tryparse_internal(DateTime, "2012-02-20T09:09:3.43i9", Dates.ISODateTimeFormat, true)
 catch err
     @test isa(err, ArgumentError)
     @test err.msg == "Found extra characters at the end of date time string"
 end
 try
-    DateTime("2012-02-20T09:09:3i439")
+    Dates.tryparse_internal(DateTime, "2012-02-20T09:09:3i439", Dates.ISODateTimeFormat, true)
 catch err
     @test isa(err, ArgumentError)
     @test err.msg == "Unable to parse date time. Expected token Delim(.) at char 19"