--- a/test/dates/io.jl
+++ b/test/dates/io.jl
@@ -68,12 +68,14 @@ b3 = "96/2/15"
 @test_throws ArgumentError Dates.DateTime(b3,f)
 try
     Dates.tryparse_internal(DateTime, "2012-02-20T09:09:3.43i9", Dates.ISODateTimeFormat, true)
+    @test false
 catch err
     @test isa(err, ArgumentError)
     @test err.msg == "Found extra characters at the end of date time string"
 end
 try
     Dates.tryparse_internal(DateTime, "2012-02-20T09:09:3i439", Dates.ISODateTimeFormat, true)
+    @test false
 catch err
     @test isa(err, ArgumentError)
     @test err.msg == "Unable to parse date time. Expected token Delim(.) at char 19"