Iterating on #18000 and #19545. Speed up is currently 140x.

before:

```julia
julia> @benchmark DateTime("2010-10-10T10:10:10.10")
BenchmarkTools.Trial: 
  memory estimate:  3.11 kb
  allocs estimate:  85
  --------------
  minimum time:     16.768 μs (0.00% GC)
  median time:      17.943 μs (0.00% GC)
  mean time:        20.198 μs (1.78% GC)
  maximum time:     3.734 ms (96.53% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

Shashi's update #19545

```julia
julia> @benchmark DateTime("2010-10-10T10:10:10.10")
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     138.344 ns (0.00% GC)
  median time:      140.283 ns (0.00% GC)
  mean time:        157.004 ns (0.00% GC)
  maximum time:     1.683 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     861
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

This PR:
```julia
julia> @benchmark DateTime("2010-10-10T10:10:10.10")
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     118.559 ns (0.00% GC)
  median time:      125.402 ns (0.00% GC)
  mean time:        138.223 ns (0.00% GC)
  maximum time:     2.867 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     907
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

fixes #15888, #13644
supersedes #18000, #19545