@omus 2016-12-07
I'm also experimenting with some ideas to apply this performance gain to all date formats. Feel free to merge in the mean time.