@omus 2016-12-14
Modified direction of the PR to work off of #19545. Mostly I am cleaning up the code without losing the speedup. I will probably also address #16928 here and may also include some custom code for specific formats.