@shashi 2016-12-15
I think this can be done by something like adding a parameter to `DatePart{char, T}` where T is the output type, in this case the type of the result in `_tryparse` can still be determined... You might have other ideas.