@omus 2016-12-15
@shashi I haven't yet dealt with parsing support for TimeZones. I have some ideas including what you suggested but I'll need to do more experimentation to find out what works best.