@omus 2016-12-16
Added fundamentals of adding custom slots to `TimeType` parsing (supports parsing of `ZonedDateTime` type). Changes don't impact performance of `Date` or `DateTime`