@shashi 2016-12-17
Nice! :)

Is ZonedDateTime parsing itself fast?