@omus 2016-12-17
ZonedDateTime parsing is about 10x faster. There are probably some additional things I can do outside of base to improve the performance more.