@omus 2016-12-19
I've added support for custom parsers which now makes DateTime parsing using ISODateTimeFormat 230x faster.

```julia
julia> @benchmark DateTime("2010-10-10T10:10:10.10")
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     70.023 ns (0.00% GC)
  median time:      77.467 ns (0.00% GC)
  mean time:        109.659 ns (0.00% GC)
  maximum time:     9.145 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     976
  time tolerance:   5.00%
  memory tolerance: 1.00%
```