@omus 2016-12-19
@shashi feel free to merge in my changes again. Code should be ready for review.