@tkelman 2016-12-19
should we change the target branch of this pr?