@omus 2016-12-19
Sure. Do I need to remove the "fixes" / "supersedes" notes before merging?