@tkelman 2016-12-19
good question, don't think so - i believe github only auto closes things that merge to master